check taxon/assembly IDs during import Fixes #30 Fixes #33
