+++ src/genomehubs/__init__.py
@@ -1,7 +1,7 @@
 """GenomeHubs."""
 
 from .genomehubs import cli
-from .lib import assembly_metadata
+from .lib import assembly
 from .lib import fill
 from .lib import init
 from .lib import taxonomy
